+++ CHANGELOG.md
@@ -5,9 +5,8 @@
 
 ### 0.11.1-dev
 
-* `lobster-json` - now all the commandline arguments except `--out` are replaced by 
-  Yaml config file which means now only `--out` and `--config` are supported command 
-  line arguments.
+* `lobster-json` - All the command-line arguments are moved to Yaml based config 
+  file. `--config` and `--out` command-line arguments are still supported.
 
 ### 0.11.0